+++ packages/manager/src/components/Chip.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Chip } from 'src/components/core/Chip';
-import type { ChipProps } from './core/Chip';
+import { Chip } from 'src/components/Chip';
+import type { ChipProps } from './Chip';
 import type { Meta, StoryObj } from '@storybook/react';
 
 export const Default: StoryObj<ChipProps> = {
@@ -62,7 +62,7 @@ export const WithDeleteButton: StoryObj<ChipProps> = {
 };
 
 const meta: Meta<ChipProps> = {
-  title: 'Components/Core/Chip',
+  title: 'Components/Chip',
   component: Chip,
   args: { label: 'Chip', onDelete: undefined },
 };
