+++ packages/manager/src/features/Kubernetes/ClusterList/KubernetesClusterRow.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Chip } from 'src/components/core/Chip';
+import { Chip } from 'src/components/Chip';
 import { Hidden } from 'src/components/Hidden';
 import Grid from '@mui/material/Unstable_Grid2';
 import { DateTimeDisplay } from 'src/components/DateTimeDisplay';
