+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/KubeSummaryPanel.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import ActionsPanel from 'src/components/ActionsPanel';
 import { Button } from 'src/components/Button/Button';
 import { ConfirmationDialog } from 'src/components/ConfirmationDialog/ConfirmationDialog';
-import { Chip } from 'src/components/core/Chip';
+import { Chip } from 'src/components/Chip';
 import Paper from 'src/components/core/Paper';
 import Grid from '@mui/material/Unstable_Grid2';
 import { TagsPanel } from 'src/components/TagsPanel/TagsPanel';
